refactor reset password email stuff to be easier to read
