+++ server/auth/index.js
@@ -4,22 +4,13 @@ const response = require('response')
 const level = require('level-party')
 const verify = require('./verify')
 const errors = require('../errors')
-const createEmail = require('township-email')
-const createReset = require('township-reset-password-token')
+const createReset = require('./reset')
 
 module.exports = function (router, db, opts) {
   const townshipDb = level(opts.township.db || path.join(__dirname, 'township.db'))
   const ship = township(townshipDb, opts.township)
-  const reset = createReset(townshipDb, {
-    secret: 'not a secret' // passed to jsonwebtoken
-  })
-
-  var email
-  if (opts.email) {
-    email = createEmail({
-      transport: opts.email.transport
-    })
-  }
+  var reset
+  if (opts.email) reset = createReset(opts.email, townshipDb)
 
   function onerror (err, res) {
     var data = {statusCode: 400, message: errors.humanize(err).message}
@@ -63,39 +54,14 @@ module.exports = function (router, db, opts) {
   })
 
   router.post('/api/v1/password-reset', function (req, res, ctx) {
-    if (!email) return onerror(new Error('config.email not set.'))
+    if (!reset) return onerror(new Error('config.email not set.'))
     const userEmail = req.body.email
     ship.accounts.findByEmail(userEmail, function (err, account) {
       if (err) return onerror(new Error('account not found'), res)
       var accountKey = account.auth.key
-      reset.create({ accountKey: accountKey }, function (err, token) {
-        if (err) return onerror(new Error('problem creating reset token'), res)
-        const clientHost = process.env.VIRTUAL_HOST
-          ? `https://${process.env.VIRTUAL_HOST}`
-          : 'http://localhost:8080'
-        var reseturl = `${clientHost}/reset-password?accountKey=${accountKey}&resetToken=${token}&email=${userEmail}`
-
-        var emailOptions = {
-          to: userEmail,
-          from: opts.email.fromEmail,
-          subject: 'Reset your password at datproject.org',
-          html: `<div>
-            <p>Hello!</p>
-            <p>You recently requested to reset your password. If that wasn't you, you can delete this email.</p>
-            <p>Reset your password by clicking this link:</p>
-            <p><b><a href="${reseturl}">Reset password</a></b></p>
-            <p>Or by following this url:</p>
-            <p><a href="${reseturl}">${reseturl}</a></p>
-          </div>`
-        }
-
-        email.send(emailOptions, function (err, info) {
-          if (err) return onerror(err, res)
-          if (opts.email.transport.name === 'Mock') {
-            console.log('mock email sent', emailOptions)
-          }
-          return response.json({ message: 'Check your email to finish resetting your password' }).pipe(res)
-        })
+      return reset.mail(userEmail, accountKey, function (err) {
+        if (err) return onerror(err, res)
+        return response.json({ message: 'Check your email to finish resetting your password' }).pipe(res)
       })
     })
   })
